+++ src/TRegx/SafeRegex/Exception/Factory/CompilePregExceptionFactory.php
@@ -56,6 +56,15 @@ class CompilePregExceptionFactory
 
     private function cleanMessage(string $message): string
     {
-        return str_replace('(PCRE2_DUPNAMES not set) ', '', $message);
+        $value = \str_replace('(PCRE2_DUPNAMES not set) ', '', $message);
+
+        if (\version_compare(PHP_VERSION, '7.3.0', '<')) {
+            if (\preg_match("/^Two named subpatterns have the same name at offset (\d+)$/", $value, $match)) {
+                $offset = $match[1] + 1; // increase offset by 1, to fix php inconsistencies
+                return "Two named subpatterns have the same name at offset $offset";
+            }
+        }
+
+        return $value;
     }
 }